Project 3b about 60 percent complete
The A star algorithm is not finding the point hence i haven't implemented the arrows as described in the homework requirement. 
Also although the map is plotting and I have a mask of the obstacle space, I haven't figured out how to use the mask to reduce the space yet. I have commented out the A star portion, I recommend running the program once first before un-commenting.